Change int to long in clustering metrics

+++ eu.mondo.map.modelmetrics/src/eu/mondo/map/modelmetrics/composite/ClusteringCoefficientList.java
@@ -18,8 +18,8 @@ public class ClusteringCoefficientList<N> extends ListMetric<Double> {
 	}
 
 	public double calculate(final Node<N> node) {
-		int interConnected = 0;
-		int numberOfNeighbors = 0;
+		long interConnected = 0;
+		long numberOfNeighbors = 0;
 		for (Node<N> neighbor1 : node.getDisjunctNeighbors()) {
 			for (Node<N> neighbor2 : node.getDisjunctNeighbors()) {
 				if (neighbor1 != neighbor2) {
